+++ src/main/java/com/myproject/service/YFinanceService.java
@@ -4,6 +4,7 @@ import com.myproject.client.YFinanceClient;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 @Slf4j
@@ -12,7 +13,7 @@ public class YFinanceService {
 
     private final YFinanceClient financeClient;
 
-    public BigInteger getPriceChange() {
+    public BigDecimal getPriceChange() {
         var price = financeClient.get10YearsPriceHistory();
         var priceChange = price.get10YearsPriceChange();
         log.info("Price change is {}%:", priceChange);
